Adding Utilities, Pikachu SageMaker Detection, restructuring code
