+++ PikachuDetection-SageMaker/Notebook/training/requirements.txt
@@ -1,0 +1,1 @@
+gluoncv==0.8.0
