+++ PikachuDetection/Lambda/PikachuDetection.py
@@ -10,7 +10,6 @@ import boto3
 
 HEIGHT = 512
 WIDTH = 512
-MODEL = 'pikachu'
 
 class PikachuDetection(panoramasdk.base):
 
@@ -19,7 +18,7 @@ class PikachuDetection(panoramasdk.base):
             "parameters":
                 (
                     ("float", "threshold", "Detection threshold", 0.50),
-                    ("model", "pokemon_detection", "Model for detecting pokemon", MODEL),
+                    ("model", "pokemon_detection", "Model for detecting pokemon", "pikachu-detection-hyb"),
                     ("int", "batch_size", "Model batch size", 1),
                     ("float", "pokemon_index", "pokemon index based on dataset used", 0),
                 ),
