--- PikachuDetection/Notebook/training/requirements.txt
@@ -1,1 +1,0 @@
-gluoncv==0.8.0
